+++ lib/Support/CMakeLists.txt
@@ -126,6 +126,7 @@ add_llvm_library(LLVMSupport
   TimeValue.cpp
   Valgrind.cpp
   Watchdog.cpp
+  WinFunctions.cpp
 
   ADDITIONAL_HEADER_DIRS
   Unix
