+++ tools/clang/tools/dxc/dxc.cpp
@@ -60,6 +60,8 @@
 #ifdef LLVM_ON_WIN32
 #include <dia2.h>
 #include <comdef.h>
+#else
+#include "llvm/Support/WinFunctions.h"
 #endif
 #include <algorithm>
 #include <unordered_map>
